Split user into user/{info,log,login,logout,router}
